breakout edgeFilter.tsx from pathfindingSearch.tsx
